Handle PUT
